+++ extensions/ark-quarkus-jackson-vertx/runtime/src/main/java/xyz/juandiii/ark/quarkus/vertx/QuarkusVertxTlsResolver.java
@@ -1,4 +1,4 @@
-package xyz.juandiii.ark.quarkus;
+package xyz.juandiii.ark.quarkus.vertx;
 
 import io.quarkus.tls.TlsConfiguration;
 import io.quarkus.tls.TlsConfigurationRegistry;