+++ extensions/ark-quarkus-jackson/deployment/src/test/java/xyz/juandiii/ark/quarkus/deployment/ArkQuarkusExtensionTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import xyz.juandiii.ark.core.ArkClient;
 import xyz.juandiii.ark.core.JsonSerializer;
-import xyz.juandiii.ark.mutiny.MutinyArkClient;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNotSame;
@@ -28,9 +27,6 @@ class ArkQuarkusExtensionTest {
     @Inject
     ArkClient.Builder anotherArkClientBuilder;
 
-    @Inject
-    MutinyArkClient.Builder mutinyArkClientBuilder;
-
     @Test
     void givenExtension_whenStarted_thenJsonSerializerIsProduced() {
         assertNotNull(jsonSerializer);
@@ -47,19 +43,8 @@ class ArkQuarkusExtensionTest {
     }
 
     @Test
-    void givenExtension_whenStarted_thenMutinyBuilderIsProduced() {
-        assertNotNull(mutinyArkClientBuilder);
-    }
-
-    @Test
     void givenSyncBuilder_whenBuild_thenCreatesClient() {
         var client = arkClientBuilder.baseUrl("http://localhost:8080").build();
-        assertNotNull(client);
-    }
-
-    @Test
-    void givenMutinyBuilder_whenBuild_thenCreatesClient() {
-        var client = mutinyArkClientBuilder.baseUrl("http://localhost:8080").build();
         assertNotNull(client);
     }
 }